Added support for ForwardRef types
Added code to evaluate typing.ForwardRef types. These are generally used to
reference types that haven't been defined yet. Unfortunately to evaluate
these types you need to pass a dictionary containing all the globals to all
the validate functions so that the type can be evaluated. This has meant
that many lines had to updated just to add the additional argument.

Also had to add special handling of the Optional type to enable the
tests to pass on Python 3.9.